Fix code format error; add commit to Image::getScaledTextureSize.

+++ src/core/images/BufferImage.h
@@ -40,7 +40,7 @@ class BufferImage : public PixelImage {
   }
 
   ISize getScaledTextureSize(float scale) const override {
-    return ISize::Make(width() * scale, height() * scale);
+    return ISize::Make(static_cast<float>(width()) * scale, static_cast<float>(height()) * scale);
   }
 
  protected:
