+++ src/core/images/CodecImage.h
@@ -41,7 +41,7 @@ class CodecImage : public GeneratorImage {
   }
 
   ISize getScaledTextureSize(float scale) const override {
-    return ISize::Make(width() * scale, height() * scale);
+    return ISize::Make(static_cast<float>(width()) * scale, static_cast<float>(height()) * scale);
   }
 
  protected:
