+++ src/core/images/PictureImage.h
@@ -49,7 +49,7 @@ class PictureImage : public Image {
   }
 
   ISize getScaledTextureSize(float scale) const override {
-    return ISize::Make(width() * scale, height() * scale);
+    return ISize::Make(static_cast<float>(width()) * scale, static_cast<float>(height()) * scale);
   }
 
   std::shared_ptr<Picture> picture = nullptr;
