+++ src/core/images/ScaledImage.h
@@ -42,7 +42,7 @@ class ScaledImage : public TransformImage {
   }
 
   ISize getScaledTextureSize(float scale) const override {
-    return ISize::Make(width() * scale, height() * scale);
+    return ISize::Make(static_cast<float>(width()) * scale, static_cast<float>(height()) * scale);
   }
 
  protected:
