+++ src/core/images/SubsetImage.h
@@ -39,7 +39,7 @@ class SubsetImage : public TransformImage {
   }
 
   ISize getScaledTextureSize(float scale) const override {
-    return ISize::Make(width() * scale, height() * scale);
+    return ISize::Make(static_cast<float>(width()) * scale, static_cast<float>(height()) * scale);
   }
 
   Rect bounds = {};
